validate on value tests in separate module
